feat: Add DTOs and mappers for Author and Book; implement services for Author, Book, Category, and Loan

+++ src/main/java/com/example/library/api/dto/request/BookRequestDTO.java
@@ -19,5 +19,5 @@ public record BookRequestDTO(
 		Integer publicationYear, 
 		Integer availableCopies,
 		Set<Long> authorIds,
-		Integer categoryId
+		Long categoryId
 ) {}
